* commit '2fb4e3e289582ff030f0d4fccc0116b8820472a2':
  bug
  bug
  bug
  bug

+++ fbi_conf/src/com/fr/bi/conf/data/source/ServerTableSource.java
@@ -147,7 +147,7 @@ public class ServerTableSource extends DBTableSource {
             return BIServerUtils.runServer(tableData, columns, new Traversal<BIDataValue>() {
                 @Override
                 public void actionPerformed(BIDataValue data) {
-                    long i = data.getRow();
+                    int i = data.getRow();
                     int j = data.getCol();
                     Object v = data.getValue();
                     Object value = null;
